wifi_nxp: fix aarch64 compile warnings
Replace uint32_t casts with uintptr_t for pointer arithmetic
to ensure correctness on 64-bit architectures. Add MLAN_64BIT
compile definition for CONFIG_ARM64.

On 64-bit (AArch64) platforms, pointers are 64-bit wide but
uint32_t is only 32-bit, causing truncation warnings when
casting pointers to integers.

- Add zephyr_library_compile_definitions_ifdef(CONFIG_ARM64
  MLAN_64BIT) in CMakeLists.txt
- Replace (uint32_t) casts with (uintptr_t) for pointer
  arithmetic
- Replace %d with %zu for sizeof() format strings
- Add OSA_HandleIsValid() helper using memcpy to safely read
  pointer-sized values

+++ mcux/middleware/wifi_nxp/port/osa/osa_zephyr.c
@@ -121,7 +121,7 @@ void *OSA_TimerGetContext(osa_timer_handle_t timerHandle)
 {
     assert(timerHandle != NULL);
 
-    struct timer_data *ptimer = (struct timer_data *)(void *)(uint32_t *)(*(uint32_t *)timerHandle);
+    struct timer_data *ptimer = (struct timer_data *)(void *)(uintptr_t *)(*(uintptr_t *)timerHandle);
 
     return ptimer->user_arg;
 }
